~ KeyManager -> CertificateStore refactoring
