+++ opc-ua-stack/stack-core/src/test/java/org/eclipse/milo/opcua/stack/core/security/MemoryCertificateStoreTest.java
@@ -1,0 +1,10 @@
+package org.eclipse.milo.opcua.stack.core.security;
+
+class MemoryCertificateStoreTest extends CertificateStoreTest {
+
+    @Override
+    protected CertificateStore newCertificateStore() {
+        return new MemoryCertificateStore();
+    }
+
+}
