--- opc-ua-stack/stack-core/src/test/java/org/eclipse/milo/opcua/stack/core/security/MemoryKeyManagerTest.java
@@ -1,10 +1,0 @@
-package org.eclipse.milo.opcua.stack.core.security;
-
-class MemoryKeyManagerTest extends KeyManagerTest {
-
-    @Override
-    protected KeyManager newKeyManager() {
-        return new MemoryKeyManager();
-    }
-
-}
